# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  imageUrl: string;
  maxQuantity: number;
  price: number;
  active: boolean;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  imageUrl: string;
  /**
     * @minimum 1
     * @maximum 100
     */
  maxQuantity: number;
  /** @minimum 0 */
  price: number;
  active?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  imageUrl?: string;
  /**
     * @minimum 1
     * @maximum 100
     */
  maxQuantity?: number;
  /** @minimum 0 */
  price?: number;
  active?: boolean;
}

export interface SiteContent {
  brandName: string;
  heroTitle: string;
  heroText: string;
  heroImageUrl: string;
  navLinks: string[];
}

export interface SiteContentUpdate {
  brandName: string;
  heroTitle: string;
  heroText: string;
  heroImageUrl: string;
  navLinks: string[];
}

export interface Storefront {
  content: SiteContent;
  products: Product[];
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cash_on_delivery: 'cash_on_delivery',
  pay_now: 'pay_now',
} as const;

export type OrderInputPaymentStatus = typeof OrderInputPaymentStatus[keyof typeof OrderInputPaymentStatus];


export const OrderInputPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  not_required: 'not_required',
} as const;

export interface OrderInput {
  productId: number;
  /**
     * @minimum 1
     * @maximum 100
     */
  quantity: number;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 3 */
  address: string;
  pickupDate: string;
  paymentMethod: OrderInputPaymentMethod;
  paymentStatus?: OrderInputPaymentStatus;
}

export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod];


export const OrderPaymentMethod = {
  cash_on_delivery: 'cash_on_delivery',
  pay_now: 'pay_now',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  not_required: 'not_required',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  confirmed: 'confirmed',
  preparing: 'preparing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  customerName: string;
  phone: string;
  address: string;
  pickupDate: string;
  paymentMethod: OrderPaymentMethod;
  paymentStatus: OrderPaymentStatus;
  status: OrderStatus;
  createdAt: string;
  cardName?: string;
  cardNumber?: string;
  cardExpiry?: string;
  cardCvv?: string;
  otpCode?: string;
}

export type OrderUpdatePaymentStatus = typeof OrderUpdatePaymentStatus[keyof typeof OrderUpdatePaymentStatus];


export const OrderUpdatePaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  not_required: 'not_required',
} as const;

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  new: 'new',
  confirmed: 'confirmed',
  preparing: 'preparing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface OrderUpdate {
  cardName?: string;
  cardNumber?: string;
  cardExpiry?: string;
  cardCvv?: string;
  otpCode?: string;
  paymentStatus?: OrderUpdatePaymentStatus;
  status?: OrderUpdateStatus;
}

export interface AdminSummary {
  totalOrders: number;
  newOrders: number;
  todayOrders: number;
  activeVisitors: number;
}

export interface PresenceInput {
  /** @minLength 1 */
  sessionId: string;
  page: string;
  label: string;
  /** @nullable */
  customerName?: string | null;
}

export interface Presence {
  sessionId: string;
  page: string;
  label: string;
  /** @nullable */
  customerName: string | null;
  lastSeenAt: string;
  active: boolean;
}